Keep `maskImageView` out of the scene proper
It is never meant to be rendered in the main panel, but is only meant to be part of the composition of final
results. Removing it from the scene is not only simpler (no need to toggle visibility), it avoids any possibility of
flickering when we introduce async preview composition.